Web scraping(python blog)/requirements.txt
psycopg2-binary
bs4

requests
html5lib==1.1